US120706 Refactor d2l-card-selection-list into separate component
